[アカウント設定] add: 退会 Promise→ try catchに変更

+++ src/components/AccountSetting/ButtonCancelMembershipSending.tsx
@@ -6,6 +6,7 @@ import { deleteUser }from 'firebase/auth'
 import { type Dispatch} from 'react'
 import { useUnsubscribe } from '../../app/UnsubscribeContext'
 import { DeleteUsers } from '../../../types/deleteUsers'
+import { FirebaseError } from "firebase/app"
 
 /** accountSettingModal.tsxから受け取るprops型を定義 */
 interface Props {
@@ -29,38 +30,53 @@ const ButtonCancelMembershipSending = (props: Props):JSX.Element => {
       createdAt: Timestamp.fromDate(new Date())
     }
 
-    // apiを使用して登録処理を行う
-    const ref = doc(db, `deleteUsers/${auth.currentUser.uid}`)
-    setDoc(ref, data)
-    .then(() =>{})
-    .catch((error) => {
-      console.log(error)
-      Alert.alert('退会ユーザー情報の登録に失敗しました。')
-    })
+    try {
+      // apiを使用して登録処理を行う
+      const ref = doc(db, `deleteUsers/${auth.currentUser.uid}`)
+      await setDoc(ref, data)
+    } catch (error: unknown) {
+      if (error instanceof FirebaseError) {
+        const { message }: { message: string } = error
+        Alert.alert(message)
+      } else {
+        Alert.alert('予期せぬエラーが発生しました\n'+error)
+      }
+    }
   }
 
   /**ユーザー削除処理 */
-  function HandledeleteUser():void {
+  async function HandledeleteUser():Promise<void> {
     // 未ログインの場合、以降の処理を実行しない
     if (!auth.currentUser){ return }
 
-    deleteUser(auth.currentUser).then(() => {
+    try {
+      await deleteUser(auth.currentUser)
       // 送信成功後、完了ダイアログを表示
       setDialogVisible(true)
-    }).catch((error) => {
-      if (!auth.currentUser){ return }
+    } catch (error: unknown) {
+      if (error instanceof FirebaseError) {
 
-      // 削除処理に失敗した場合、ログイン中ユーザーのuidを指定して削除ユーザー履歴を削除
-      deleteDoc(doc(db, `deleteUsers/` , auth.currentUser.uid))
-      const { code, message }: { code: string, message: string } = error
+        if (!auth.currentUser){ return }
 
-      // 最後にログインから長期間経過
-      if(code === 'auth/requires-recent-login' ){
-        Alert.alert('長期間ログインされていない為、再ログイン後に操作を行って下さい。')
-        return
+        try {
+          // 削除処理に失敗した場合、ログイン中ユーザーのuidを指定して削除ユーザー履歴を削除
+          await deleteDoc(doc(db, `deleteUsers/` , auth.currentUser.uid))
+        } catch (deleteError) {
+          Alert.alert('予期せぬエラーが発生しました\n'+deleteError)
+        }
+
+        const { message }: { message: string } = error
+
+        // 最後にログインから長期間経過
+        if(message === 'auth/requires-recent-login' ){
+          Alert.alert('長期間ログインされていない為、再ログイン後に操作を行って下さい。')
+          return
+        }
+        Alert.alert(message)
+      } else {
+        Alert.alert('予期せぬエラーが発生しました\n'+error)
       }
-      Alert.alert(message)
-    })
+    }
   }
 
   // 退会処理送信
@@ -68,18 +84,27 @@ const ButtonCancelMembershipSending = (props: Props):JSX.Element => {
     // 未ログインの場合、以降の処理を実行しない
     if (!auth.currentUser){ return }
 
-    // 削除ユーザー履歴登録処理
-    await regCancelMember()
+    try {
+      // 削除ユーザー履歴登録処理
+      await regCancelMember()
 
-    // 削除ユーザー履歴登録結果確認
-    const docRef = doc(db, `deleteUsers/`, auth.currentUser.uid)
-    const result = await getDoc(docRef)
+      // 削除ユーザー履歴登録結果確認
+      const docRef = doc(db, `deleteUsers/`, auth.currentUser.uid)
+      const result = await getDoc(docRef)
 
-    // 削除ユーザー履歴の登録実行※削除ユーザー履歴が登録されていない場合は実行しない
-    // 削除ユーザー履歴登録失敗のエラーはregCancelMember()関数が表示
-    if(result.exists()){
-      if (unsubscribe) unsubscribe()
-      HandledeleteUser()
+      // 削除ユーザー履歴の登録実行※削除ユーザー履歴が登録されていない場合は実行しない
+      // 削除ユーザー履歴登録失敗のエラーはregCancelMember()関数が表示
+      if(result.exists()){
+        if (unsubscribe) unsubscribe()
+        await HandledeleteUser()
+      }
+    } catch (error: unknown) {
+      if (error instanceof FirebaseError) {
+        const { message }: { message: string } = error
+        Alert.alert(message)
+      } else {
+        Alert.alert('予期せぬエラーが発生しました\n'+error)
+      }
     }
   }
